requirements.txt
python-telegram-bot>=21.0
google-api-python-client>=2.116.0
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
python-dotenv>=1.0.0
groq>=0.4.0
pydantic>=2.0
httpx>=0.25.0
aiohttp>=3.9.0